app/requirements.txt
streamlit
azure-storage-blob
openai
pandas
requests
newspaper3k
retry
streamlit-folium
azure-cosmos
wordcloud